+++ spring-boot-x-demo/src/main/java/orgg/openingo/x/controller/EsController.java
@@ -61,7 +61,7 @@ import java.util.Map;
 /**
  * EsController
  *
- * @author zhucongqi
+ * @author Qicz
  */
 @RestController
 @RequestMapping("/es")