+++ src/main/java/org/openingo/spring/boot/SpringBootVersionX.java
@@ -27,6 +27,9 @@
 
 package org.openingo.spring.boot;
 
+/**
+ * @author Qicz
+ */
 public final class SpringBootVersionX {
 
 	private SpringBootVersionX() {