+++ RELEASENOTES.md
@@ -1,5 +1,32 @@
 ## Release Notes
 
+### 2.9.RELEASE
+
+- copy SpringBoot configs
+
+```java
+/**
+ * App
+ *
+ * @author Qicz
+ */
+@SpringBootApplication
+@EnableExtension
+public class App {
+
+    public static void main(String[] args) throws InterruptedException {
+        SpringApplicationX.run(App.class, args);
+        SpringApplicationX.applicationInfo();
+    }
+}
+```
+
+- use args `ccp` just copy application configs
+
+```bash
+java -jar xxx.jar ccp
+```
+
 ### 2.5.RELEASE
 
 - support Elasticsearch `MappingsProperties` configuration.